+++ ceammc/ext/src/conv/conv_ms2str.cpp
@@ -1,5 +1,5 @@
 /*****************************************************************************
- * Copyright 2017 Serge Poltavsky. All rights reserved.
+ * Copyright 2023 Serge Poltavsky. All rights reserved.
  *
  * This file may be distributed under the terms of GNU Public License version
  * 3 (GPL v3) as defined by the Free Software Foundation (FSF). A copy of the
@@ -11,12 +11,12 @@
  * contact the author of this file, or the owner of the project in which
  * this file belongs to.
  *****************************************************************************/
-#include "conv_sec2str.h"
+#include "conv_ms2str.h"
 #include "ceammc_convert.h"
 #include "ceammc_factory.h"
 #include "datatype_string.h"
 
-SecToStr::SecToStr(const PdArgs& a)
+MsToStr::MsToStr(const PdArgs& a)
     : BaseObject(a)
     , ms_flag_(nullptr)
     , symbol_out_(nullptr)
@@ -29,9 +29,9 @@ SecToStr::SecToStr(const PdArgs& a)
     addProperty(symbol_out_);
 }
 
-void SecToStr::onFloat(t_float v)
+void MsToStr::onFloat(t_float v)
 {
-    const std::string str = convert::time::sec2str(v, ms_flag_->value()).c_str();
+    const std::string str = convert::time::sec2str(v / 1000, ms_flag_->value()).c_str();
 
     if (symbol_out_->value())
         symbolTo(0, gensym(str.c_str()));
@@ -39,13 +39,13 @@ void SecToStr::onFloat(t_float v)
         atomTo(0, new DataTypeString(str));
 }
 
-void setup_conv_sec2str()
+void setup_conv_ms2str()
 {
-    ObjectFactory<SecToStr> obj("conv.sec2str");
-    obj.addAlias("sec->str");
+    ObjectFactory<MsToStr> obj("conv.ms2str");
+    obj.addAlias("ms->str");
     obj.useDefaultPdListFn();
 
-    obj.setDescription("converts float time in seconds to formated symbol (or string)");
+    obj.setDescription("converts float time in milliseconds to formated symbol (or string)");
     obj.setCategory("conv");
     obj.setKeywords({"conv", "time"});
 }