+++ ceammc/ext/src/conv/conv_ms2str.h
@@ -1,0 +1,32 @@
+/*****************************************************************************
+ * Copyright 2023 Serge Poltavsky. All rights reserved.
+ *
+ * This file may be distributed under the terms of GNU Public License version
+ * 3 (GPL v3) as defined by the Free Software Foundation (FSF). A copy of the
+ * license should have been included with this file, or the project in which
+ * this file belongs to. You may also find the details of GPL v3 at:
+ * http://www.gnu.org/licenses/gpl-3.0.txt
+ *
+ * If you have any questions regarding the use of this file, feel free to
+ * contact the author of this file, or the owner of the project in which
+ * this file belongs to.
+ *****************************************************************************/
+#ifndef TIME_MS2STR_H
+#define TIME_MS2STR_H
+
+#include "ceammc_object.h"
+
+using namespace ceammc;
+
+class MsToStr : public BaseObject {
+    FlagProperty* ms_flag_;
+    FlagProperty* symbol_out_;
+
+public:
+    MsToStr(const PdArgs& a);
+    void onFloat(t_float v) override;
+};
+
+void setup_conv_ms2str();
+
+#endif // TIME_MS2STR_H