+++ ceammc/ext/src/conv/mod_conv.cpp
@@ -19,6 +19,7 @@ void setup_conv_bpm2ms();
 void setup_conv_bpm2sec();
 void setup_conv_ms2bpm();
 void setup_conv_ms2samp();
+void setup_conv_ms2str();
 void setup_conv_samp2ms();
 void setup_conv_samp2sec();
 void setup_conv_sec2bpm();
@@ -61,6 +62,7 @@ void ceammc_conv_setup()
     setup_conv_bpm2sec();
     setup_conv_ms2bpm();
     setup_conv_ms2samp();
+    setup_conv_ms2str();
     setup_conv_samp2ms();
     setup_conv_samp2sec();
     setup_conv_sec2bpm();